Move deleteItem action to Item screen
