+++ admin/client/App/screens/Item/actions.js
@@ -3,6 +3,7 @@ import {
 	LOAD_DATA,
 	DATA_LOADING_SUCCESS,
 	DATA_LOADING_ERROR,
+	DELETE_ITEM,
 } from './constants';
 
 /**
@@ -60,5 +61,35 @@ export function dataLoadingError (error) {
 	return {
 		type: DATA_LOADING_ERROR,
 		error,
+	};
+}
+
+/**
+ * Deletes an item and optionally redirects to the current list URL
+ *
+ * @param  {String} id     The ID of the item we want to delete
+ * @param  {Object} router A react-router router object. If this is passed, we
+ *                         redirect to Keystone.adminPath/currentList.path!
+ */
+export function deleteItem (id, router) {
+	return (dispatch, getState) => {
+		const list = getState().lists.currentList;
+		list.deleteItem(id, (err) => {
+			// If a router is passed, redirect to the current list path,
+			// otherwise stay where we are
+			if (router) {
+				router.push(`${Keystone.adminPath}/${list.path}`);
+			}
+			// TODO Proper error handling
+			if (err) {
+				alert('Error deleting item, please try again!');
+				console.log(err);
+			} else {
+				dispatch({
+					type: DELETE_ITEM,
+					id,
+				});
+			}
+		});
 	};
 }
