+++ admin/client/App/screens/Item/constants.js
@@ -2,3 +2,4 @@ export const SELECT_ITEM = 'app/Item/SELECT_ITEM';
 export const LOAD_DATA = 'app/Item/LOAD_DATA';
 export const DATA_LOADING_SUCCESS = 'app/Item/DATA_LOADING_SUCCESS';
 export const DATA_LOADING_ERROR = 'app/Item/DATA_LOADING_ERROR';
+export const DELETE_ITEM = 'app/List/DELETE_ITEM';
